refactor(SDK-2575): update init method name with initialize
